\name{comparisonAllStructuresData1}
\docType{data}
\alias{comparisonAllStructuresData1}
\title{Result of regression on all 64 possible structures for Data set 1}
\description{Result of nonlinear regression on all 64 possible network structures 
for a two species system generated by networkStructures() and run over observation 
from the two species system exampleData1}
\usage{data(comparisonAllStructuresData1)}
\format{List}
\keyword{datasets}

