\name{comparisonTwoStructuresData1}
\docType{data}
\alias{comparisonTwoStructuresData1}
\title{Result of regression on two possible structures for Data set 1}
\description{Result of nonlinear regression on two network structures
for a two species system generated by networkStructures() and run over observation
from the two species system exampleData1}
\usage{data(comparisonTwoStructuresData1)}
\format{List}
\keyword{datasets}

